GRAPH CENTRALITY METRICS
Centrality metrics are used to identify the most important nodes within a graph. There are various centrality metrics including Degree, Closeness, Betweenness, Eigenvector, Katz, PageRank, etc. In the scope of this project, you are expected to implement the following two centrality metrics and test on the given two graph datasets.
1.Betweenness Centrality
2.Closeness Centrality